+++ src/MicroElements.Metadata/Validation/Rules/IRequiredPropertyValidationRule.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) MicroElements. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace MicroElements.Validation.Rules
+{
+    /// <summary>
+    /// Marker interface for <see cref="Required{T}"/>.
+    /// </summary>
+    public interface IRequiredPropertyValidationRule : IPropertyValidationRule
+    {
+    }
+}